refactor: Player라는 추상클래스 생성후 참가자와 딜러 클래스가 상속받도록 수정
