+++ src/main/java/model/Dealer.java
@@ -1,8 +1,8 @@
 package model;
 
-public class Dealer extends Participant{
+public class Dealer extends Player{
     public Dealer() {
-        super("딜러");
+        this.name = "딜러";
     }
 
     public boolean isEnough() {
